+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - (Feature) JobScheduler Volumes, Probes, Lifecycle and Ports integration
 - (Feature) Merge ArangoDB Usage Metrics
 - (Bugfix) Check Connection to the ArangoDB before creating Backup
+- (Feature) Deployment & Members Condition metrics
 
 ## [1.2.38](https://github.com/arangodb/kube-arangodb/tree/1.2.38) (2024-02-22)
 - (Feature) Extract GRPC Server